refactor(css): move concatenate export generation into css module generator
