+++ crates/rspack_plugin_css/src/utils.rs
@@ -8,21 +8,10 @@ use std::{
 use cow_utils::CowUtils;
 use heck::{ToKebabCase, ToLowerCamelCase};
 use regex::{Captures, Regex};
-use rspack_core::{
-  ChunkGraph, CompilerOptions, CssExport, CssExportsConvention, GenerateContext, LocalIdentName,
-  PathData, RESERVED_IDENTIFIER, ResourceData, RuntimeGlobals, UsedNameItem,
-  rspack_sources::{ConcatSource, RawStringSource},
-  to_identifier,
-};
+use rspack_core::{CompilerOptions, CssExportsConvention, LocalIdentName, PathData, ResourceData};
 use rspack_error::{Diagnostic, Error, Result, Severity};
 use rspack_hash::{HashDigest, HashFunction, HashSalt, RspackHash};
-use rspack_util::{
-  atom::Atom,
-  fx_hash::{FxIndexMap, FxIndexSet},
-  identifier::make_paths_relative,
-  itoa, json_stringify_str,
-};
-use rustc_hash::FxHashSet as HashSet;
+use rspack_util::identifier::make_paths_relative;
 
 pub const AUTO_PUBLIC_PATH_PLACEHOLDER: &str = "__RSPACK_PLUGIN_CSS_AUTO_PUBLIC_PATH__";
 pub static LEADING_DIGIT_REGEX: LazyLock<Regex> =
@@ -193,111 +182,6 @@ pub(crate) fn export_locals_convention(
     res.push(key.to_kebab_case());
   }
   res
-}
-
-pub fn css_modules_exports_to_concatenate_module_string<'a>(
-  exports: FxIndexMap<&'a str, &'a FxIndexSet<CssExport>>,
-  module: &dyn rspack_core::Module,
-  generate_context: &mut GenerateContext,
-  concate_source: &mut ConcatSource,
-) -> Result<()> {
-  let GenerateContext {
-    compilation,
-    concatenation_scope,
-    runtime,
-    runtime_template,
-    ..
-  } = generate_context;
-  let Some(scope) = concatenation_scope else {
-    return Ok(());
-  };
-  let module_graph = compilation.get_module_graph();
-  let mut used_identifiers = HashSet::default();
-  let exports_info = compilation
-    .exports_info_artifact
-    .get_exports_info_data(&module.identifier());
-  for (key, elements) in exports {
-    let export_info = exports_info.get_read_only_export_info(&Atom::from(key));
-    let used_name = export_info.get_used_name(None, *runtime);
-    let used_name = match used_name {
-      Some(UsedNameItem::Str(name)) => name.to_string(),
-      _ => key.to_string(),
-    };
-
-    let content = elements
-      .iter()
-      .map(
-        |CssExport {
-           ident,
-           from,
-           id: _,
-           orig_name: _,
-         }| match from {
-          None => json_stringify_str(ident),
-          Some(from_name) => {
-            let from = module
-              .get_dependencies()
-              .iter()
-              .find_map(|id| {
-                let dependency = module_graph.dependency_by_id(id);
-                let request = if let Some(d) = dependency.as_module_dependency() {
-                  Some(d.request())
-                } else {
-                  dependency.as_context_dependency().map(|d| d.request())
-                };
-                if let Some(request) = request
-                  && request == from_name
-                {
-                  return module_graph.module_graph_module_by_dependency_id(id);
-                }
-                None
-              })
-              .expect("should have css from module");
-
-            let from_exports_info = compilation
-              .exports_info_artifact
-              .get_exports_info_data(&from.module_identifier);
-            let from_used_name = match from_exports_info
-              .get_read_only_export_info(&Atom::from(ident.as_str()))
-              .get_used_name(None, *runtime)
-            {
-              Some(UsedNameItem::Str(name)) => json_stringify_str(&name),
-              _ => json_stringify_str(ident),
-            };
-
-            let from = rspack_util::json_stringify(
-              ChunkGraph::get_module_id(&compilation.module_ids_artifact, from.module_identifier)
-                .expect("should have module"),
-            );
-            format!(
-              "{}({from})[{}]",
-              runtime_template.render_runtime_globals(&RuntimeGlobals::REQUIRE),
-              from_used_name
-            )
-          }
-        },
-      )
-      .collect::<Vec<_>>()
-      .join(" + \" \" + ");
-    let mut identifier: Cow<'_, str> = Cow::Owned(to_identifier(&used_name).into_owned());
-    if RESERVED_IDENTIFIER.contains(identifier.as_ref()) {
-      identifier = Cow::Owned(format!("_{identifier}"));
-    }
-    let mut i = 0;
-    while used_identifiers.contains(&identifier) {
-      let mut i_buffer = itoa::Buffer::new();
-      let i_str = i_buffer.format(i);
-      identifier = Cow::Owned(format!("{identifier}{i_str}"));
-      i += 1;
-    }
-    // TODO: conditional support `const or var` after we finished runtimeTemplate utils
-    concate_source.add(RawStringSource::from(format!(
-      "var {identifier} = {content};\n"
-    )));
-    used_identifiers.insert(identifier.clone());
-    scope.register_export(key.into(), identifier.into_owned());
-  }
-  Ok(())
 }
 
 static STRING_MULTILINE: LazyLock<Regex> =
